fixed devql config max lines
